requirements.txt
accelerate==1.11.0
alembic==1.17.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
appnope==0.1.4
asttokens==3.0.0
blinker==1.9.0
cachetools==6.2.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cloudpickle==3.1.2
colorlog==6.10.1
comm==0.2.3
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
databricks-sdk==0.73.0
debugpy==1.8.17
decorator==5.2.1
docker==7.1.0
executing==2.2.1
fastapi==0.121.2
filelock==3.20.0
Flask==3.1.2
flask-cors==6.0.1
fonttools==4.60.1
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.43.0
graphene==3.4.3
graphql-core==3.2.7
graphql-relay==3.2.0
gunicorn==23.0.0
h11==0.16.0
hf-xet==1.2.0
huey==2.5.4
huggingface-hub==0.36.0
idna==3.11
importlib_metadata==8.7.0
ipykernel==7.1.0
ipython==9.6.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jupyter_client==8.6.3
jupyter_core==5.9.1
kiwisolver==1.4.9
littleutils==0.2.4
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.5
numpy==2.3.4
ogb==1.3.6
opentelemetry-api==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
optuna==4.6.0
outdated==0.2.2
packaging==25.0
pandas==2.3.3
parso==0.8.5
peft==0.17.1
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.0
prompt_toolkit==3.0.52
protobuf==6.33.1
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
regex==2025.11.3
requests==2.32.5
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.44
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.49.3
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.9.0
torchinfo==1.8.0
torchvision==0.24.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
wcwidth==0.2.14
Werkzeug==3.1.3
wilds==2.0.0
zipp==3.23.0
accelerate==1.11.0
alembic==1.17.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
appnope==0.1.4
asttokens==3.0.0
beautifulsoup4==4.14.2
blinker==1.9.0
cachetools==6.2.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cloudpickle==3.1.2
colorlog==6.10.1
comm==0.2.3
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
databricks-sdk==0.73.0
debugpy==1.8.17
decorator==5.2.1
dnspython==2.8.0
docker==7.1.0
email-validator==2.3.0
executing==2.2.1
fastapi==0.121.2
fastapi-cli==0.0.20
fastapi-cloud-cli==0.8.0
fastar==0.8.0
filelock==3.20.0
Flask==3.1.2
flask-cors==6.0.1
fonttools==4.60.1
fsspec==2025.10.0
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.43.0
graphene==3.4.3
graphql-core==3.2.7
graphql-relay==3.2.0
gunicorn==23.0.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huey==2.5.4
huggingface-hub==0.36.0
idna==3.11
ImageIO==2.37.2
importlib_metadata==8.7.0
ipykernel==7.1.0
ipython==9.6.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jupyter_client==8.6.3
jupyter_core==5.9.1
kiwisolver==1.4.9
lazy_loader==0.4
littleutils==0.2.4
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
mdurl==0.1.2
mlflow==3.6.0
mlflow-skinny==3.6.0
mlflow-tracing==3.6.0
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.5
numpy==2.2.6
ogb==1.3.6
opencv-python==4.12.0.88
opentelemetry-api==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
optuna==4.6.0
outdated==0.2.2
packaging==25.0
pandas==2.3.3
parso==0.8.5
peft==0.17.1
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.0
polars==1.35.2
polars-runtime-32==1.35.2
prompt_toolkit==3.0.52
protobuf==6.33.1
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.2.5
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.21
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
regex==2025.11.3
requests==2.32.5
rich==14.2.0
rich-toolkit==0.17.1
rignore==0.7.6
rsa==4.9.1
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
segment-anything==1.0
sentry-sdk==2.48.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.49.3
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2025.10.16
tokenizers==0.22.1
torch==2.9.0
torchinfo==1.8.0
torchvision==0.24.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
typer==0.21.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.3.231
ultralytics-thop==2.0.18
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
wcwidth==0.2.14
websockets==15.0.1
Werkzeug==3.1.3
wilds==2.0.0
zipp==3.23.0
